+++ lib/eva/client/remote_item.py
@@ -38,6 +38,7 @@ class RemoteUpdatableItem(eva.item.UpdatableItem):
         self._virtual_allowed = False
         self._snmp_traps_allowed = False
         self._drivers_allowed = False
+        self._modbus_allowed = False
         cfg = {}
         if controller.mqtt_update:
             cfg['mqtt_update'] = controller.mqtt_update