+++ lib/eva/lm/lvar.py
@@ -18,6 +18,7 @@ class LVar(eva.item.VariableItem):
         self._virtual_allowed = False
         self._snmp_traps_allowed = False
         self._drivers_allowed = False
+        self._modbus_allowed = False
         self.status = 1
         self.mqtt_update_topics.append('set_time')
         self.prv_value = None